requirements.txt
aniso8601==9.0.1
attrs==22.1.0
autobahn==19.11.2
Automat==20.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==38.0.1
enum34==1.1.10
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiteconnect==4.1.0
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.1
pandas-ta==0.3.14b0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
pyotp==2.7.0
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
service-identity==21.1.0
six==1.16.0
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
urllib3==1.26.12
Werkzeug==2.2.2
zope.interface==5.5.0
